Support external environments for commands
this adds `conda project run --external-environment=<env>`

Where `<env>` can be a relative or absolute path to a conda environment,
or the name of a global environment stored in one of `env_dirs` directories.

If there is a collision between a local environment directory and named
environment, the local environment directory is preferred.

If no local environment directory found it will check through the conda
config"env_dirs" in order and return the first matching named
environment.

In this PR a defined environment is not required.